[CHANGE] refactor VocabularyListTile
